pyproject.toml
[project]
name = "intelligent-analyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "fastmcp>=2.11.3",
    "google-genai>=1.31.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "mcp-server-time>=2025.8.4",
    "mcp[cli]>=1.13.1",
    "python-dotenv>=1.1.1",
]
